Move widget click key from note widget light to widget utils
