+++ Simplenote/src/main/java/com/automattic/simplenote/NoteEditorActivity.java
@@ -27,10 +27,10 @@ import org.wordpress.passcodelock.AppLockManager;
 
 import java.util.ArrayList;
 
-import static com.automattic.simplenote.NoteWidgetLight.KEY_WIDGET_CLICK;
 import static com.automattic.simplenote.analytics.AnalyticsTracker.CATEGORY_WIDGET;
 import static com.automattic.simplenote.analytics.AnalyticsTracker.Stat.NOTE_WIDGET_NOTE_TAPPED;
 import static com.automattic.simplenote.utils.DisplayUtils.disableScreenshotsIfLocked;
+import static com.automattic.simplenote.utils.WidgetUtils.KEY_WIDGET_CLICK;
 
 public class NoteEditorActivity extends AppCompatActivity {
     private TabLayout mTabLayout;
